Add GroupByEq iterator.
This iterator groups the two input streams based on whether
corresponding items are equal.  It returns whether they are equal and
the length until that comparison changes value.

We have also extended its behaviour in the following ways to aid its
use in the compression routines:

- it is often useful to add an initial "Same(0)" group, corresponding
to the initial skip operation; and

- it is also useful to ignore final Same operation, since there's no
need to store it due to the op count byte.