Add GroupBySS2 iterator.
SS2 is slightly weird in that skips work in bytes, whereas memcpy and
memsets are work in shorts.  If the run ends with a Diff group, but
there is an odd number of bytes, then that final group will be split
into an even length Diff group followed by a length 1 Diff group.